+++ src/main/java/springbook/user/dao/CountingDaoFactory.java
@@ -2,6 +2,8 @@ package springbook.user.dao;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import springbook.user.dao.connection.CountingConnectionMaker;
+import springbook.user.dao.connection.SimpleConnectionMaker;
 
 @Configuration
 public class CountingDaoFactory {
